chore: working on threshold

+++ src/consensus/mod.rs
@@ -3,7 +3,7 @@ pub mod block_production;
 use serde::Deserialize;
 use std::default::Default;
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Clone)]
 pub struct EpochRandomness(pub String);
 
 impl Default for EpochRandomness {
@@ -12,9 +12,9 @@ impl Default for EpochRandomness {
     }
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Clone)]
 pub struct Epoch {
-    pub c: u32,
+    pub c: f64,
     pub epoch_length: u32,
     pub randomness: EpochRandomness,
 }
@@ -22,14 +22,14 @@ pub struct Epoch {
 impl Default for Epoch {
     fn default() -> Self {
         Epoch {
-            c: 0,
+            c: 0.0,
             epoch_length: 10,
             randomness: Default::default(),
         }
     }
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Clone)]
 pub struct ProductionAuthorities(pub Vec<String>);
 
 impl Default for ProductionAuthorities {
